cleanup and login required component restyled
